Move edge tests
